changed index to size_t

+++ src/display/DisplayList.cpp
@@ -48,6 +48,6 @@ string DisplayList::get_input() const {
     return ConsoleInput().read_line();
 }
 
-std::optional<const Item *> DisplayList::select_child_at(int i) {
+std::optional<const Item *> DisplayList::select_child_at(size_t i) {
     return nullopt;
 }
